Cloudrun-Agents/capital-agent-docker/requirements.txt
google-adk
fastapi
uvicorn